Update Digit Truncation to Rounding
At the moment, if more digits than the normally allowed maximum are specified, additional digits will be truncated away rather than being rounded. This may seem small, but for those writing automated traders based on this library, it might be significant after many trades. Added in a round() before converting the value to a string and truncating. Please merge if you feel appropriate.